Run Cucumber tests in a testthat context

It's purpose is to be able to run Cucumber tests alongside testthat tests.

To do that, place a call to run() in one of the test-*.R files in your tests/testthat directory.

Usage

run(path = ".", filter = NULL, ...)

Arguments

path

Path to the directory containing the .feature files. If run() is placed in a tests/testthat/test-*.R file and you call testthat::test_dir or similar, it runs in the tests/testthat directory. The default value "." finds all feature files in the tests/testthat directory.

filter

If not NULL, only features with file names matching this regular expression will be executed. Matching is performed on the file name after it's stripped of ".feature".

...

Additional arguments passed to grepl().

Value

NULL, invisibly. To get result and a report, use cucumber::test(), or inspect the result of testthat function call.

Examples

if (FALSE) { # \dontrun{
#' tests/testthat/test-cucumber.R
cucumber::run()
} # }