Run all Cucumber tests

This command runs all Cucumber tests. It takes all .feature files from the features_dir and runs them using the steps from the steps_dir.

Usage

test(
  features_dir,
  steps_dir,
  steps_loader = .default_steps_loader,
  test_interactive = getOption("cucumber.interactive", default = FALSE)
)

Arguments

features_dir

A character string of the directory containing the feature files.

steps_dir

A character string of the directory containing the step files.

steps_loader

A function that loads the steps implementations. By default it sources all files from the steps_dir using the built-in mechanism. You can provide your own function to load the steps. The function should take one argument, which will be the steps_dir.

For packages: set to NULL to use default support-code load mechanism and place your steps definitions in setup or helper files. Read more about those files in testthat documentation.

test_interactive

A logical value indicating whether to ask which feature files to run.

Value

None, function called for side effects.

Examples

if (FALSE) { # \dontrun{
#' testthat/acceptance/test-cucumber.R
cucumber::test(".", ".", steps_loader = NULL)
# Steps are stored in `testthat/acceptance/setup-steps.R` file.
} # }